Fixed class cast error

+++ src/test/java/project/query/ArenaObjectRingSelectorTest.java
@@ -42,8 +42,8 @@ public class ArenaObjectRingSelectorTest extends ArenaObjectSelectorTest {
             randomParams[i] = new Object[] {
                 RANDOM_X_COOR.get(),
                 RANDOM_Y_COOR.get(),
-                Math.min(r1, r2),
-                Math.max(r1, r2)
+                (short) Math.min(r1, r2),
+                (short) Math.max(r1, r2)
             };
         }
